Revert "[ObsAssistant] Add Insights component to alerts details (#178330)"
This reverts commit 8feb50dd467115148b02569e7d85e57d601e3ede.
