+++ packages/kbn-es-types/src/search.ts
@@ -421,7 +421,6 @@ export type AggregateOf<
           {
             doc_count: number;
             key: string[];
-            key_as_string: string;
           } & SubAggregateOf<TAggregationContainer, TDocument>
         >;
       };
@@ -561,16 +560,6 @@ export type AggregateOf<
             doc_count: number;
             key: string | number;
             key_as_string?: string;
-          } & SubAggregateOf<TAggregationContainer, TDocument>
-        >;
-      };
-      categorize_text: {
-        buckets: Array<
-          {
-            doc_count: number;
-            key: string | number;
-            regex?: string;
-            max_matching_length: number;
           } & SubAggregateOf<TAggregationContainer, TDocument>
         >;
       };
