+++ x-pack/plugins/observability_solution/apm/public/components/alerting/ui_components/alert_details_app_section/index.tsx
@@ -22,12 +22,7 @@ import { useKibana } from '@kbn/kibana-react-plugin/public';
 import { getPaddedAlertTimeRange } from '@kbn/observability-get-padded-alert-time-range-util';
 import { EuiCallOut } from '@elastic/eui';
 import { CoreStart } from '@kbn/core/public';
-import {
-  SERVICE_ENVIRONMENT,
-  SERVICE_NAME,
-  TRANSACTION_NAME,
-  TRANSACTION_TYPE,
-} from '../../../../../common/es_fields/apm';
+import { SERVICE_ENVIRONMENT } from '../../../../../common/es_fields/apm';
 import { ChartPointerEventContextProvider } from '../../../../context/chart_pointer_event/chart_pointer_event_context';
 import { TimeRangeMetadataContextProvider } from '../../../../context/time_range_metadata/time_range_metadata_context';
 import { getComparisonChartTheme } from '../../../shared/time_comparison/get_comparison_chart_theme';
@@ -36,9 +31,13 @@ import { getAggsTypeFromRule } from './helpers';
 import { LatencyAlertsHistoryChart } from './latency_alerts_history_chart';
 import LatencyChart from './latency_chart';
 import ThroughputChart from './throughput_chart';
-import { AlertDetailsAppSectionProps } from './types';
+import {
+  AlertDetailsAppSectionProps,
+  SERVICE_NAME,
+  TRANSACTION_TYPE,
+  TRANSACTION_NAME,
+} from './types';
 import { createCallApmApi } from '../../../../services/rest/create_call_apm_api';
-import { AlertDetailContextualInsights } from './alert_details_contextual_insights';
 
 export function AlertDetailsAppSection({
   rule,
@@ -52,7 +51,6 @@ export function AlertDetailsAppSection({
   const alertRuleTypeId = alert.fields[ALERT_RULE_TYPE_ID];
   const alertEvaluationValue = alert.fields[ALERT_EVALUATION_VALUE];
   const alertEvaluationThreshold = alert.fields[ALERT_EVALUATION_THRESHOLD];
-
   const environment = alert.fields[SERVICE_ENVIRONMENT];
   const serviceName = String(alert.fields[SERVICE_NAME]);
   const transactionName = alert.fields[TRANSACTION_NAME];
@@ -102,19 +100,6 @@ export function AlertDetailsAppSection({
         ),
         value: serviceName,
       },
-      ...(transactionName
-        ? [
-            {
-              label: (
-                <FormattedMessage
-                  id="xpack.apm.pages.alertDetails.alertSummary.transactionName"
-                  defaultMessage="Transaction name"
-                />
-              ),
-              value: transactionName,
-            },
-          ]
-        : []),
     ];
     setAlertSummaryFields(alertSummaryFields);
   }, [
@@ -123,7 +108,6 @@ export function AlertDetailsAppSection({
     alertEvaluationThreshold,
     environment,
     serviceName,
-    transactionName,
     setAlertSummaryFields,
   ]);
 
@@ -166,8 +150,6 @@ export function AlertDetailsAppSection({
 
   return (
     <EuiFlexGroup direction="column" gutterSize="s">
-      <AlertDetailContextualInsights alert={alert} />
-
       <TimeRangeMetadataContextProvider
         start={from}
         end={to}
