+++ x-pack/plugins/observability_solution/apm/public/components/alerting/ui_components/alert_details_app_section/types.ts
@@ -11,14 +11,11 @@ import type {
   AlertSummaryField,
 } from '@kbn/observability-plugin/public';
 import type { TIME_UNITS } from '@kbn/triggers-actions-ui-plugin/public';
-import type {
-  SERVICE_NAME,
-  TRANSACTION_TYPE,
-  SERVICE_ENVIRONMENT,
-  TRANSACTION_NAME,
-  CONTAINER_ID,
-  HOST_NAME,
-} from '../../../../../common/es_fields/apm';
+import type { SERVICE_ENVIRONMENT } from '../../../../../common/es_fields/apm';
+
+export const SERVICE_NAME = 'service.name' as const;
+export const TRANSACTION_TYPE = 'transaction.type' as const;
+export const TRANSACTION_NAME = 'transaction.name' as const;
 export interface AlertDetailsAppSectionProps {
   rule: Rule<{
     environment: string;
@@ -29,10 +26,8 @@ export interface AlertDetailsAppSectionProps {
   alert: TopAlert<{
     [SERVICE_NAME]: string;
     [TRANSACTION_TYPE]: string;
-    [CONTAINER_ID]?: string;
-    [HOST_NAME]?: string;
-    [TRANSACTION_NAME]?: string;
     [SERVICE_ENVIRONMENT]: string;
+    [TRANSACTION_NAME]: string;
   }>;
   timeZone: string;
   setAlertSummaryFields: React.Dispatch<
