--- x-pack/plugins/observability_solution/apm/public/context/kibana_context/use_kibana.tsx
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { CoreStart } from '@kbn/core/public';
-import { useKibana } from '@kbn/kibana-react-plugin/public';
-import { type ApmPluginStartDeps } from '../../plugin';
-
-export type StartServices<AdditionalServices extends object = {}> = CoreStart &
-  ApmPluginStartDeps &
-  AdditionalServices;
-
-const useTypedKibana = <AdditionalServices extends object = {}>() =>
-  useKibana<StartServices<AdditionalServices>>();
-
-export { useTypedKibana as useKibana };
