+++ x-pack/plugins/observability_solution/apm/server/assistant_functions/get_apm_downstream_dependencies.ts
@@ -18,7 +18,7 @@ export function registerGetApmDownstreamDependenciesFunction({
       name: 'get_apm_downstream_dependencies',
       contexts: ['apm'],
       description: `Get the downstream dependencies (services or uninstrumented backends) for a 
-      service. This allows you to map the downstream dependency name to a service, by 
+      service. This allows you to map the dowstream dependency name to a service, by 
       returning both span.destination.service.resource and service.name. Use this to 
       drilldown further if needed.`,
       descriptionForUser: i18n.translate(
