+++ x-pack/plugins/observability_solution/apm/server/lib/helpers/get_random_sampler/index.ts
@@ -9,8 +9,6 @@ import { KibanaRequest } from '@kbn/core/server';
 import seedrandom from 'seedrandom';
 import { APMRouteHandlerResources } from '../../../routes/apm_routes/register_apm_server_routes';
 
-export type RandomSampler = Awaited<ReturnType<typeof getRandomSampler>>;
-
 export async function getRandomSampler({
   security,
   request,
@@ -37,3 +35,5 @@ export async function getRandomSampler({
     seed,
   };
 }
+
+export type RandomSampler = Awaited<ReturnType<typeof getRandomSampler>>;
