+++ x-pack/plugins/observability_solution/apm/server/routes/alerts/register_apm_rule_types.ts
@@ -23,10 +23,8 @@ import { legacyExperimentalFieldMap } from '@kbn/alerts-as-data-utils';
 import type { APMIndices } from '@kbn/apm-data-access-plugin/server';
 import {
   AGENT_NAME,
-  CONTAINER_ID,
   ERROR_GROUP_ID,
   ERROR_GROUP_NAME,
-  HOST_NAME,
   PROCESSOR_EVENT,
   SERVICE_ENVIRONMENT,
   SERVICE_LANGUAGE_NAME,
@@ -49,14 +47,6 @@ export const apmRuleTypeAlertFieldMap = {
     required: false,
   },
   [SERVICE_ENVIRONMENT]: {
-    type: 'keyword',
-    required: false,
-  },
-  [HOST_NAME]: {
-    type: 'keyword',
-    required: false,
-  },
-  [CONTAINER_ID]: {
     type: 'keyword',
     required: false,
   },
