+++ x-pack/plugins/observability_solution/apm/server/routes/alerts/rule_types/get_service_group_fields.test.ts
@@ -6,7 +6,7 @@
  */
 
 import {
-  getApmAlertSourceFields,
+  getServiceGroupFields,
   getServiceGroupFieldsAgg,
   flattenSourceDoc,
 } from './get_service_group_fields';
@@ -18,12 +18,6 @@ const mockSourceObj = {
     language: {
       name: 'typescript',
     },
-  },
-  container: {
-    id: 'my-container',
-  },
-  host: {
-    name: 'my-host',
   },
   labels: {
     team: 'test',
@@ -44,12 +38,10 @@ const mockBucket = {
 
 describe('getSourceFields', () => {
   it('should return a flattened record of fields and values for a given bucket except for labels', () => {
-    const result = getApmAlertSourceFields(mockBucket);
+    const result = getServiceGroupFields(mockBucket);
     expect(result).toMatchInlineSnapshot(`
       Object {
         "agent.name": "nodejs",
-        "container.id": "my-container",
-        "host.name": "my-host",
         "labels": Object {
           "event": Array [
             "event-0",
@@ -79,8 +71,6 @@ describe('getSourceFieldsAgg', () => {
                 "service.environment",
                 "service.language.name",
                 "labels",
-                "host.name",
-                "container.id",
               ],
             },
             "size": 1,
@@ -105,8 +95,6 @@ describe('getSourceFieldsAgg', () => {
                 "service.environment",
                 "service.language.name",
                 "labels",
-                "host.name",
-                "container.id",
               ],
             },
             "size": 1,
@@ -130,8 +118,6 @@ describe('flattenSourceDoc', () => {
     expect(result).toMatchInlineSnapshot(`
       Object {
         "agent.name": "nodejs",
-        "container.id": "my-container",
-        "host.name": "my-host",
         "labels": Object {
           "event": Array [
             "event-0",
