+++ x-pack/plugins/observability_solution/apm/server/routes/alerts/rule_types/get_service_group_fields.ts
@@ -6,17 +6,10 @@
  */
 
 import { AggregationsTopHitsAggregation } from '@elastic/elasticsearch/lib/api/types';
-import { CONTAINER_ID, HOST_NAME } from '../../../../common/es_fields/apm';
 import {
   LABELS,
   SERVICE_GROUP_SUPPORTED_FIELDS,
 } from '../../../../common/service_groups';
-
-const APM_ALERT_SOURCE_FIELDS = [
-  ...SERVICE_GROUP_SUPPORTED_FIELDS,
-  HOST_NAME,
-  CONTAINER_ID,
-];
 
 export interface SourceDoc {
   [key: string]: string | string[] | SourceDoc;
@@ -30,7 +23,7 @@ export function getServiceGroupFieldsAgg(
       top_hits: {
         size: 1,
         _source: {
-          includes: APM_ALERT_SOURCE_FIELDS,
+          includes: SERVICE_GROUP_SUPPORTED_FIELDS,
         },
         ...topHitsOpts,
       },
@@ -46,7 +39,7 @@ interface AggResultBucket {
   };
 }
 
-export function getApmAlertSourceFields(bucket?: AggResultBucket) {
+export function getServiceGroupFields(bucket?: AggResultBucket) {
   if (!bucket) {
     return {};
   }
