+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/get_apm_timeseries/get_error_event_rate.ts
@@ -6,7 +6,6 @@
  */
 
 import type { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
-import { ERROR_GROUP_NAME } from '../../../../common/es_fields/apm';
 import { ApmDocumentType } from '../../../../common/document_type';
 import { RollupInterval } from '../../../../common/rollup';
 import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
@@ -41,14 +40,7 @@ export async function getErrorEventRate({
       rollupInterval: RollupInterval.None,
       intervalString,
       filter,
-      groupByFields: [ERROR_GROUP_NAME],
       aggs: {
-        sample: {
-          top_metrics: {
-            metrics: { field: ERROR_GROUP_NAME },
-            sort: '_score',
-          },
-        },
         value: {
           bucket_script: {
             buckets_path: {
