+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/get_apm_timeseries/get_exit_span_failure_rate.ts
@@ -50,7 +50,7 @@ export async function getExitSpanFailureRate({
           spanDestinationServiceResource
         )
       ),
-      groupByFields: [SPAN_DESTINATION_SERVICE_RESOURCE],
+      groupBy: SPAN_DESTINATION_SERVICE_RESOURCE,
       aggs: {
         successful: {
           filter: {
