+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/get_apm_timeseries/get_exit_span_throughput.ts
@@ -49,7 +49,7 @@ export async function getExitSpanThroughput({
           spanDestinationServiceResource
         )
       ),
-      groupByFields: [SPAN_DESTINATION_SERVICE_RESOURCE],
+      groupBy: SPAN_DESTINATION_SERVICE_RESOURCE,
       aggs: {
         value: {
           bucket_script: {
