+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/get_apm_timeseries/get_transaction_throughput.ts
@@ -8,10 +8,7 @@
 import type { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
 import { termQuery } from '@kbn/observability-plugin/server';
 import { ApmDocumentType } from '../../../../common/document_type';
-import {
-  TRANSACTION_NAME,
-  TRANSACTION_TYPE,
-} from '../../../../common/es_fields/apm';
+import { TRANSACTION_TYPE } from '../../../../common/es_fields/apm';
 import { RollupInterval } from '../../../../common/rollup';
 import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
 import { fetchSeries } from './fetch_timeseries';
@@ -24,7 +21,6 @@ export async function getTransactionThroughput({
   bucketSize,
   filter,
   transactionType,
-  transactionName,
 }: {
   apmEventClient: APMEventClient;
   start: number;
@@ -33,7 +29,6 @@ export async function getTransactionThroughput({
   bucketSize: number;
   filter: QueryDslQueryContainer[];
   transactionType?: string;
-  transactionName?: string;
 }) {
   const bucketSizeInMinutes = bucketSize / 60;
   const rangeInMinutes = (end - start) / 1000 / 60;
@@ -48,12 +43,8 @@ export async function getTransactionThroughput({
       documentType: ApmDocumentType.TransactionMetric,
       rollupInterval: RollupInterval.OneMinute,
       intervalString,
-      filter: [
-        ...filter,
-        ...termQuery(TRANSACTION_TYPE, transactionType),
-        ...termQuery(TRANSACTION_NAME, transactionName),
-      ],
-      groupByFields: [TRANSACTION_TYPE, TRANSACTION_NAME],
+      filter: filter.concat(...termQuery(TRANSACTION_TYPE, transactionType)),
+      groupBy: 'transaction.type',
       aggs: {
         value: {
           bucket_script: {
