+++ x-pack/plugins/observability_solution/apm/server/routes/assistant_functions/route.ts
@@ -10,13 +10,7 @@ import { omit } from 'lodash';
 import { getApmAlertsClient } from '../../lib/helpers/get_apm_alerts_client';
 import { getApmEventClient } from '../../lib/helpers/get_apm_event_client';
 import { getMlClient } from '../../lib/helpers/get_ml_client';
-import { getRandomSampler } from '../../lib/helpers/get_random_sampler';
 import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
-import {
-  apmAlertDetailsContextRt,
-  getApmAlertDetailsContext,
-} from './get_apm_alert_details_context';
-
 import {
   CorrelationValue,
   correlationValuesRouteRt,
@@ -32,78 +26,11 @@ import {
   serviceSummaryRouteRt,
   type ServiceSummary,
 } from './get_apm_service_summary';
-import { ApmAnomalies } from './get_apm_service_summary/get_anomalies';
 import {
   getApmTimeseries,
   getApmTimeseriesRt,
-  TimeseriesChangePoint,
   type ApmTimeseries,
 } from './get_apm_timeseries';
-import { LogCategories } from './get_log_categories';
-
-const getApmAlertDetailsContextRoute = createApmServerRoute({
-  endpoint: 'GET /internal/apm/assistant/get_apm_alert_details_context',
-  options: {
-    tags: ['access:apm', 'access:ai_assistant'],
-  },
-
-  params: t.type({
-    query: apmAlertDetailsContextRt,
-  }),
-  handler: async (
-    resources
-  ): Promise<{
-    serviceSummary: ServiceSummary;
-    downstreamDependencies: APMDownstreamDependency[];
-    logCategories: LogCategories;
-    serviceChangePoints: Array<{
-      title: string;
-      changes: TimeseriesChangePoint[];
-    }>;
-    exitSpanChangePoints: Array<{
-      title: string;
-      changes: TimeseriesChangePoint[];
-    }>;
-    anomalies: ApmAnomalies;
-  }> => {
-    const { context, request, plugins, logger, params } = resources;
-    const { query } = params;
-
-    const alertStartedAt = query.alert_started_at;
-
-    const [
-      apmEventClient,
-      annotationsClient,
-      coreContext,
-      apmAlertsClient,
-      mlClient,
-    ] = await Promise.all([
-      getApmEventClient(resources),
-      plugins.observability.setup.getScopedAnnotationsClient(context, request),
-      context.core,
-      getApmAlertsClient(resources),
-      getMlClient(resources),
-      getRandomSampler({
-        security: resources.plugins.security,
-        probability: 1,
-        request: resources.request,
-      }),
-    ]);
-    const esClient = coreContext.elasticsearch.client.asCurrentUser;
-
-    return getApmAlertDetailsContext({
-      coreContext,
-      alertStartedAt,
-      annotationsClient,
-      apmAlertsClient,
-      apmEventClient,
-      esClient,
-      logger,
-      mlClient,
-      query,
-    });
-  },
-});
 
 const getApmTimeSeriesRoute = createApmServerRoute({
   endpoint: 'POST /internal/apm/assistant/get_apm_timeseries',
@@ -233,7 +160,6 @@ const getApmCorrelationValuesRoute = createApmServerRoute({
 
 export const assistantRouteRepository = {
   ...getApmTimeSeriesRoute,
-  ...getApmAlertDetailsContextRoute,
   ...getApmServiceSummaryRoute,
   ...getApmCorrelationValuesRoute,
   ...getDownstreamDependenciesRoute,
